requirements.txt
asgiref==3.11.0
channels==4.3.2
channels_redis==4.3.0
crispy-bootstrap5==2023.10
cryptography==45.0.4
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.58.2
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-filter==23.5
django-leaflet==0.33.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
html2docx==1.6.0
lxml==6.0.2
mammoth==1.11.0
openpyxl==3.1.2
pdf2image==1.17.0
pillow==10.2.0
PyJWT==2.10.1
python-decouple==3.8
python-docx==1.2.0
python3-openid==3.2.0
pytz==2025.2
pywin32==311
redis==7.1.0
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
tinycss2==1.5.1
tzdata==2025.2
webencodings==0.5.1
